[feat] 미용실 등록한 게시글 조회 API 연결
