+++ front/src/app/(main)/salon/[salonName]/[salonAddress]/_component/Header.tsx
