+++ front/src/app/(main)/salon/[salonName]/[salonAddress]/page.tsx
@@ -5,13 +5,17 @@ import SalonPosts from './_component/SalonPosts';
 import { getSalonPosts } from './_lib/getSalonPosts';
 
 type Props = {
-  params: { salonName: string};
+  params: { 
+    salonName: string,
+    salonAddress: string
+  };
 }
 
 export default async function Page({params}: Props) {
-  const {salonName} = params;
+  const salonName = decodeURIComponent(params.salonName);
+  const salonAddress = decodeURIComponent(params.salonAddress);
   const queryClient = new QueryClient();
-  await queryClient.prefetchQuery({queryKey: ['salon', salonName], queryFn: getSalonPosts});
+  await queryClient.prefetchQuery({queryKey: ['salon', salonName, salonAddress], queryFn: getSalonPosts});
   const dehydrateState = dehydrate(queryClient);
   return (
     <>
@@ -19,7 +23,7 @@ export default async function Page({params}: Props) {
         <HydrationBoundary state={dehydrateState}>
           <Header salonName={salonName}/>
           <div className={style.postsWrapper}>
-            <SalonPosts salonName={salonName}/>
+            <SalonPosts salonName={salonName} salonAddress={salonAddress}/>
           </div>
         </HydrationBoundary>
       </div>
